requirements.txt
loguru~=0.7.2
aiogram~=3.3.0
aiogram_dialog~=2.1.0
tortoise-orm~=0.20.0
redis~=5.0.1
asyncpg~=0.29.0
aerich~=0.7.2a